🔧 add the configuration for HTTP v1
The configuration that needs to be specified for v1 calls is different from the
legacy FCM API
https://github.com/olucurious/PyFCM?tab=readme-ov-file#updates-breaking-changes

We now need to specify the path to the service account JSON file, and the
project_id. Adding these to the config file temporarily so that we can test it
on staging more easily. Once they are known to work, we can put them into the
list of environment variables for the cloud services team to configure.