Task A-4: Run image_build_push only after tests pass successfully
Approach:
Converted test-with-docker and test-with-manual-install to reusable workflows.
Added them as jobs in image_build_push workflow.
Build job in this workflow needs these test jobs as dependencies.

-------

Test Workflow run
https://github.com/MukuFlash03/e-mission-server/actions/runs/11096277197/job/30825929886

-----

Notes

Github actions didn't have out of the box solution for running a workflow based on results of multiple workflows where ALL workflows must have completed successfully.
We need this since both the test-with-docker and test-with-manual-install must pass.
So this needs an "AND" logic.

"workflow_run" is there but this triggers the dependent workflow when either of the workflow dependencies defined as prerequisites are completed.
So this has an "OR" logic.
https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/events-that-trigger-workflows#workflow_run

----

Found an alternative suggestion here:
https://stackoverflow.com/a/75597437

This suggests converting the pre-requisite workflows into reusable workflows.
These workflows can then be called in jobs in the workflow that needs these workflows to be run before.
Finally, these jobs can be added as dependencies for the requisite job.

In our scenario, two new jobs are added to the image_build_push.yml for each of the two tests environments.
These will run parallelly.

Then in the build image job, these jobs are added in the "needs" field, indicating that these jobs must pass successfully before running the build job.

-------

Also corrected the branch in reusable workflow for fetching latest server image tag.
Need to update in the currently open PRs as well.

------

Also, removed the `push` trigger from the two `test` workflows since the `image_build_push` workflow would also be triggered on a `push` event which in its workflow triggers these two test workflows.
Thus if not removed, each test workflow would run twice.