Revert "Modified userstat where it passes in the trip_key as an input instead of hardcoding analysis/confirmed_trip"
This reverts commit c77c5c549d1839dbcb56b0921126e769edfdebe6.

Reverting because this broke the tests.
https://github.com/e-mission/e-mission-server/pull/1006#issuecomment-2573807472

This is because the composite trips include both actual trips and untracked time.
https://github.com/e-mission/e-mission-server/pull/1006#issuecomment-2574122106

We don't want to store untracked time in the user stats or show it to users, so
we can't push this to a real production system, where people check the admin
dashboard, just yet. The fix is simple, but will take @TeachMeTW some time to
figure out. So let's revert the change for now.