Revert "Always run the computation of the user profile stats"
This reverts commit 9d1f414b067cbb8411214fed88d45aaa8c6d5eae.

Always running the computation of the user profile stats meant that we were
retrieving all composite trips every hour. This could have possibly led to a
significantly greater I/O load, since DocumentDB appears to be very slow for
composite trip queries.

However, we cannot just change the pipeline range to use confirmed trips
instead of composite either, since the UI relies on composite trips. If the
pipeline failed after creating confirmed trips but before creating composite
trips, and we use composite trips for the pipeline range, the user will not see
the trips in the UI, since we only pull draft trips for the time range after
the pipeline is done https://github.com/e-mission/e-mission-server/pull/1011

The real fix is to split the user stats into pipeline-related and non-pipeline-related.
I will make that change in a subsequent commit in the same PR