🐛 Do not iterate over map keys while modifying them
In https://github.com/e-mission/e-mission-server/pull/1008 (and in particular, https://github.com/e-mission/e-mission-server/commit/1ec706e7e0d51cee3dd5942f87b45f3a9b9dd86d), we munged incoming keys with dots so that we could store them properly in mongodb.

We did this by iterating over all the keys, "popping" the one with dots and
storing the value into the munged key, consistent with https://stackoverflow.com/a/11277439

However, this lead to an issue where apparently the key order is also changed,
so that modified keys are iterated over twice, and some keys are not even
checked. And since they are not checked, they are not munged.

This refactors the code to two stages:
- find the `keys_to_munge` (aka `ktm`) by iterating over the map
- iterate over the `keys_to_munge` list and munge the keys

Testing done:

- Before this, the tests in e94140ea55a78c201d3f4c10d4833841769d4cac were failing

```
----------------------------------------------------------------------
Ran 8 tests in 28.153s

FAILED (failures=2, errors=1)
```

- After this, the tests in e94141ea55a78c201d3f4c10d4833841769d4cac were passing

```
----------------------------------------------------------------------
Ran 8 tests in 28.969s

OK
```