⚡️ Update the call-specific metrics inline instead of querying the database
We have made several changes to cache summary information in the user profile.
This summary information can be used to improve the scalability of the admin
dashboard (https://github.com/e-mission/op-admin-dashboard/issues/145) but will
also be used to determine dormant deployment and potentially in a future
OpenPATH-wide dashboard.

These changes started by making a single call to cache both trip and call stats
https://github.com/e-mission/e-mission-server/pull/1005

This resulted in all the composite trips being read every hour, so we split the
stats into pipeline-dependent and pipeline-independent stats, in
88bb35a79fda4026842e1b59d2e2e85a73893b09
(part of https://github.com/e-mission/e-mission-server/pull/1005)

The idea was that, since the composite object query was slow because the
composite trips were large, we could run only the queries for the server API
stats every hour, and read the composite trips only when they were updated.

However, after the improvements to the trip segmentation pipeline
(https://github.com/e-mission/e-mission-server/pull/1014, results in
https://github.com/e-mission/e-mission-docs/issues/1105#issuecomment-2631914594)
reading the server stats is now the bottleneck.

Checking the computation time on the big deployments (e.g. ccebikes), although
the time taken has significantly improved as the database load has gone down,
even in the past two days, we see a median of ~ 10 seconds and a max of over
two minutes.

And we don't really need to query this data to generate the call summary
statistics. Instead of computing them on every run, we can compute them only
when _they_ change, which happens when we receive calls to the API.

So, in the API's after_request hook, in addition to adding a new stat, we can
also just update the `last_call_ts` in the profile. This potentially makes
every API call a teeny bit slower since we are adding one more DB write, but it
significantly lowers the DB load, so should make the system as a whole faster.

Testing done:
- the modified `TestUserStat` test for the `last_call_ts` passes