👀 Add additional call summaries to the profile
`last_call_ts` is necessary but not sufficient. While it is helpful to know
when the phone last contacted the server, we really want to know whether the
call was in the foreground or the background, whether the call involved an
upload, and whether the upload included any sensor (location) data. We should
cache all this in the profile as well.

This change adds the following entries:
- the last sync (get or put)
- the last put
- the last diary refresh

It also adds three new tests, one to check for each of the new precomputed calls.

It also refactors the user stat test slightly since we now have one test that
checks for trips, and four tests that check for user calls. The checks for user
calls don't need any trips to work, so the steps in `setUp` to load the data
and run the intake pipeline is wasted. So we just pull out the trip loading
steps into a separate function and invoke it only from
`testGetAndStoreUserStats`.

I tried to make it a decorator (that is why it is named `testWithTrips`), but
ran into an issue with passing in arguments correctly.  Running out of time to
make code beautiful, so just calling the new function manually for now instead.

Testing done:

```
$ ./e-mission-py.bash emission/tests/analysisTests/intakeTests/TestUserStat.py
----------------------------------------------------------------------
Ran 6 tests in 9.480s

OK
```