👀 Add upload data to the profile
This change builds on 05c8e91ec126022433cc31784362ee8bca75e5b1 to store basic
stats about the uploads from the phone to the profile.

The two stats here are:
- the last uploaded location
- the last uploaded phone data

These are also updated directly from the usercache `phone_to_server` function,
which is the main location where the incoming data is stored into the e-mission
database, albeit in the usercache.

We also update the tests for the builtin usercache to check for these fields
being filled in. After every call to `phone_to_server`, we check that the
profile fields have been filled in properly.

Note that, because these fields are updated only when new data comes in, they
will not work for all users, particularly the ones that have stopped
contributing data. We need to create a historical migration script to fill
those in for dormant users.

Testing done:

```
$ ./e-mission-py.bash emission/tests/netTests/TestBuiltinUserCache.py
----------------------------------------------------------------------
Ran 9 tests in 0.397s

OK
```