👀 Store both create and update timestamps in the profile
Before this change, in `/profile/create`, and by implication, in
`User.register`, we would insert new values using `upsert`. So we had only one
modification timestamp, which was `update_ts`. However, this meant that if a
user uninstalled + reinstalled, or logged out and logged in, the `update_ts`
would be changed, and we would lose the information of when the user
registered.

In this change, on creation, we set both create and update timestamps to the
same value. When the user re-registers, we only change the update_ts on the
existing entry, thus allowing us to see both the initial install and the most
recent one.

We insert entries into two databases - the UUID DB and the profile DB, and
follow the same check-and-insert pattern with create and insert timestamps for
both.

This also refactors the implementation of the registration functionality to
follow a consistent pattern and avoid the confusion from the `register` ->
`registerWithUUID` -> `_createInitialProfile` mapping.

This is also a fix for https://github.com/e-mission/e-mission-docs/issues/1111#issuecomment-2654695639

Testing done:

Before change:

After Creation

```
[{'_id': ObjectId('67acf82e46756f8244f43d3b')
 'user_id': UUID('d4ed8ea7-fdc5-4e0b-b03d-f68ea99a0327')
 'mode': {}
 'mpg_array': [32.044384997811726]
 'purpose': {}
 'source': 'Shankari'
 'update_ts': datetime.datetime(2025, 2, 12, 11, 36, 14, 657000)}]
```

After logout + login

```
[{'_id': ObjectId('67acf82e46756f8244f43d3b')
 'user_id': UUID('d4ed8ea7-fdc5-4e0b-b03d-f68ea99a0327')
 'mode': {}
 'mpg_array': [32.044384997811726]
 'purpose': {}
 'source': 'Shankari'
 'update_ts': datetime.datetime(2025, 2, 12, 11, 36, 14, 657000)
 'client_app_version': '1.9.4'
 'client_os_version': '18.1'
 'curr_platform': 'ios'
 'manufacturer': 'Apple'
 'phone_lang': 'en'
 'last_call_ts': 1739389168.35327
 'last_diary_fetch_ts': 1739389168.305688}]
```

After change:

After creation

```
{'_id': ObjectId('67ad36113d56a7ca523171ef')
 'user_id': UUID('1dd58310-311f-40fd-bb51-5fa1e3a2ecbe')
 'source': 'Shankari'
 'create_ts': datetime.datetime(2025, 2, 12, 16, 0, 17, 368000)
 'update_ts': datetime.datetime(2025, 2, 12, 16, 0, 17, 368000)
 'mpg_array': [32.044384997811726]
 'mode': {}
 'purpose': {}
 'last_call_ts': 1739404817.396993}
```

After logout + login

```
{'_id': ObjectId('67ad36113d56a7ca523171ef')
 'user_id': UUID('1dd58310-311f-40fd-bb51-5fa1e3a2ecbe')
 'source': 'Shankari'
 'create_ts': datetime.datetime(2025, 2, 12, 16, 0, 17, 368000)
 'update_ts': datetime.datetime(2025, 2, 12, 16, 2, 59, 284000)
 'mpg_array': [32.044384997811726]
 'mode': {}
 'purpose': {}
 'last_call_ts': 1739404979.297653}
```