⚡️ Specify the appropriate hints on the indices to use for various queries
While looking into an upgrade to mongo 8.0.4, we found that the queries were
super slow. So running the pipeline after reset for a specified set of three
users took ~ 15 mins against mongo 4.4.0 but over a day against mongo 8.0.4
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2676415132

Investigating further, including by looking at the slow queries, we saw that
this was due to the `metadata.key` index being used for several of the queries.
Using the `metadata.key` instead of `data.ts` index for location queries, for
example, took 3x the time.
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2677074300

I think that the medium-term fix is to use a composite index. However, we can
also specify a "hint" while running the query, which will encourage mongodb to
use the correct index.
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2677093069

However, the `data.ts` hint only seems to work if used with a `data.ts` filter
in the query - e.g. for `background/location`. While reading the `manual/*`
entries, the filter query works with `metadata.write_ts`, and then, querying by
both `data.ts` and `metadata.write_ts` is slow. The only index that is
moderately fast is `metadata.key`. As an aside, I assume that is why the
default index was set to `metadata.key` in the first place; the query plan was
optimized for the first set of queries and then subsequent calls with the same
"shape" used the same index.
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2677474876

Note that I still don't understand why `metadata.write_ts` is slow for the
`manual/*` series.

Regardless, a fairly simple fix which says

> But for now, we can just say that if the sort key is `metadata.write_ts` we can
> make the hint be `metadata.key`.

With this fix, the run against mongodb 8.0.4 is also < 15 mins
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2677583319

Testing done:
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2677583319