🗃️ Recreate indices after dropping collections during testing
Without this change, all indices were dropped when the collections were dropped.
This would show up as a performance hit earlier. But now, as part of
b23ffb97fd22fb4084a63b3f0e461d28f976b4c5, we add an index hint, so the query
fails. Instead of taking the performance hit, we recreate the index so that our
queries will work and will work fast ⚡️⚡️⚡️

I had originally implemented this by reloading the module, but that ran into
issues with the inheritance chain
https://github.com/e-mission/e-mission-server/pull/1028#issuecomment-2679741539

Refactoring the initialization of the map into a function and using it to
reinitialize the map works.

This fixes
https://github.com/e-mission/e-mission-server/pull/1028#issuecomment-2679621529

Testing done:

```
----------------------------------------------------------------------
Ran 373 tests in 785.793s

FAILED (failures=2)
```

The two failures are because I am running locally with a non standard DB name
(`openpath_prod_ccebikes`)