💩 Have the hints be optional so we can run on DocDB without errors
In https://github.com/e-mission/e-mission-server/pull/1028 I added some hints
to indicate which indices MongoDB should use for specific queries. When
deploying this to staging/production, however, it turns out that DocDB does not
support hints on sparse indices.
https://github.com/e-mission/e-mission-docs/issues/1109#issuecomment-2686341450

If we attempt to specify a hint on a sparse index, it is not just ignored, but
the operation fails. So this is a blocker for testing out any other changes
(unrelated to hints) on staging/production.

To unblock testing against real data while exploring a migration from DocDB to
MongoDB, we need to remove hints on DocDB and only on DocDB. I could simply
revert the previous change , but that would distort our experimentation with
scalability improvements on dev against MongoDB

So I instead plumb through a option (`USE_HINTS`), defaulting to false, that
controls this. While running against MongoDB in a dev environment, we should
set this to True. Since we will not change the environment variables on
production, we will not use hints there.

All the related changes are in this single commit so it can be reverted later.

Testing done:
- Ran the pipeline against the USAID dataset. No related errors.
- Set USE_HINTS=True and then re-ran the pipeline against the USAID dataset.
  No related errors.
- Set USE_HINTS=True in one of the testing workflows (test-with-manual-install)
  but leave it as False for the docker testing so we can test both cases.