have timeseries `find_entries` return entries as list, not iterable
pymongo queries return a Cursor, which we had been converting to an iterator before returning it from `find_entries`. To get the count of entries, we were making a separate `count_documents` query.
In most places where `find_entries` is called, we were consuming the iterator into a list, so we might as well do that upfront in `_get_entries_for_timeseries`. That way, we can use the `len` of the list rather than making a second `count_documents` query.

The main changes are in builtin_timeseries; the rest of the changes are just removing the numerous `list(...)` conversions in places where `find_entries` is called, as well as renaming variables to reflect that `find_entries` returns a list, not an iterator

> # This list() conversion causes the cursor to be consumed, which is memory-expensive,
> # but faster as it allows us to get the count without having to make an additional
> "count_documents" DB call.
> # In a future situation where memory is constrained or DB calls are less expensive,
> # we could skip the list() conversion and return the cursor directly.
> # https://github.com/e-mission/e-mission-server/pull/1032#issuecomment-2685846702