🔥 Remove the pykov dependency
We had originally introduced pykov as a dependency because it was used the
markov-based tour models. However, we never integrated the `user_model_josh`
into the system, and while we integrated the tour model into common places and
trips, we have since replaced it with the trip model.

$ grep -r pykov emission/
emission//analysis/modelling/user_model_josh/utility_model.py:import pykov as pk
emission//storage/decorations/common_place_queries.py:import pykov as pk

`pykov` is no longer available through pip
https://github.com/e-mission/e-mission-server/pull/1032#issuecomment-2692461749

So this can be a forcing function to finally delete it.

To fix the usages, we also delete the entire `user_model_josh` directory since
it has never been used
And comment out the one one reference to it in `common_place_queries`

That one reference is in `get_successor`, which is only used in
`emission//incomplete_tests/TestCommonPlaceQueries.py` so it should not cause
any tests to fail.

Testing done:
- Automated tests pass (checked via CI)