⚡️ Use previously-read transition entries instead of re-reading them
While testing the improvements in
https://github.com/e-mission/e-mission-server/pull/1017

I noticed that there were multiple calls to `get_data_df` during
TRIP_SEGMENTATION, even though we should be reading all the values upfront.

https://github.com/e-mission/e-mission-server/pull/1017#issuecomment-2693016450
https://github.com/e-mission/e-mission-server/pull/1017#issuecomment-2693018236

On further investigation, this turned out to be from the
`is_tracking_restarted_in_range` function, which is used to determine whether
we need to add untracked time entries or not.
https://github.com/e-mission/e-mission-server/pull/1017#issuecomment-2693018236

This is O(p), so it is not very intensive, but it is also redundant since we
start off by reading all transitions in the active time range into memory.

We tried to use the newly added `tracking_restarted_in_loc_df` but it only has
comparisons with the previous point. We tried to see if we could determine the
range to consider and then check for `all()` but the `tracking_restarted` flag is not
available outside of the dist/time subclasses.
https://github.com/e-mission/e-mission-server/pull/1017#issuecomment-2693070905
https://github.com/e-mission/e-mission-server/pull/1017#issuecomment-2693097262

So we plumbed the transition_df through the codebase and used it directly.

`is_tracking_restarted` used to take a timeseries or a transition_df, we have
replaced it with the transition_df since we don't want to read from the
database any more.

Testing done:
 - all "real data" unit tests pass locally

```
$ ./e-mission-py.bash emission/tests/analysisTests/intakeTests/TestPipelineRealData.py
----------------------------------------------------------------------
Ran 36 tests in 159.777s

OK
```