add MONITOR_DB option to record stats for all commands to DB
For local development, we have been using this QueryMonitor to identify spots in the pipeline with excessive DB calls. For every command to the DB, this records an entry with the type of command and the list of functions in the current stack trace (to identify where the DB call originated from)

It is only enabled when MONITOR_DB environment variable is set. We will not enable this on production since it causes thousands of extra entries to be recorded to the DB