🔊 Add additional logs to showcase the multi out-of-order case
In https://github.com/e-mission/e-mission-docs/issues/1122 we found that
although we filter out out of order (OOO) trips already, sometimes, there are
two successive points that are out of order. So even after the existing code
dropps the OOO points, we end up with residual OOO points which cause errors
downstream, notably by creating a trip with a negative duration - aka end
before start.

This allows us to understand the multi-OOO case better so we can figure out how
to fix it.