add '_id' as a secondary sort key for timeseries queries
MongoDB sort queries produce inconsistent results when sorting on a field that has duplicate values. According to mongoDB docs, an easy way to prevent this is adding '_id' in the sort query, as '_id' is guaranteed to be unique

This resolves the issue being discussed in https://github.com/e-mission/e-mission-server/pull/1042#issuecomment-2779249958
Tests pass when following those steps.