🧪 Add a locust test suite to simulate a phone user
The proportion of the calls is roughly the same as the top 4 types of calls on production.

```
/profile/update                        1287
/usercache/get                          581
/usercache/put                          260
/datastreams/find_entries/timestamp      93
/pipeline/get_range_ts                   24
/profile/get                             22
/profile/create                           5
/result/metrics/yyyy_mm_dd                2
```

We currently have an adaptive wait time but it should be simple to change that
to a fixed one. This expects that there is an existing database, configured
using DB_HOST that can be the source of the data to put to the server.

Testing done:
Ran the suite and saw the variation in the requests per second.