🔧 Remove the top level `http` directive
While deploying a similar nginx profile on production, we got the error

```
2025-04-18T18:57:49.108-07:00 2025/04/19 01:57:49 [emerg] 1#0: "http" directive is not allowed here in /opt/nginx/conf.d/default.conf:1
2025-04-18T18:57:49.108-07:00 nginx: [emerg] "http" directive is not allowed here in /opt/nginx/conf.d/default.conf:1
```

The problem is that this file is not the top level `nginx.conf`, but a
sub-level `/opt/nginx/conf.d/default.conf`

If we mount the original config at `/etc/nginx/conf.d/default.conf` (as seen in
the change to the docker-compose) we get the same error. The fix is just to
remove the `http` directive - when this is included, the new lines are
automatically put in the directive and work.

Testing done: Ran with the changes and verified that the rate limiting

```
2025/04/19 16:31:20 [error] 21#21: *1038 limiting requests, excess: 2.523 by zone "profile_update", client: 192.168.2.3, server: , request: "POST /api/profile/update HTTP/1.1", host: "nginxrp"
2025/04/19 16:31:24 [error] 21#21: *1038 limiting requests, excess: 2.300 by zone "profile_update", client: 192.168.2.3, server: , request: "POST /api/profile/update HTTP/1.1", host: "nginxrp"
```